Reference problem solved
quote: 'In VBA, data elements in user-defined data types are packed to 4-byte boundaries, whereas in Visual Studio, by default, they are packed to 8-byte boundaries. Therefore you must enclose the C/C++ structure definition in a #pragma pack(4) … #pragma pack() block to avoid elements being misaligned.

+++ ELL/ELL/TestPassVariables.cpp
@@ -1,5 +1,6 @@
 #include "stdafx.h"
 #include "TestPassVariables.h"
+#include <iostream>
 
 int __stdcall ellPassStruct(typTmp &tmp) {
 	int r;
@@ -20,6 +21,7 @@ void __stdcall ellPassStruct03(typTmp &tmp) {
 }
 
 void __stdcall ellChangeRemotely(typTmp &tmp) {
-	tmp.c += tmp.a + tmp.b ; //tmp.c++;
+	//if (tmp.c == 4) { tmp.b = 5; }
+	tmp.c += (double)tmp.a + (double)tmp.b; //tmp.c++;
 }
